requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.7
cssselect==1.2.0
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
et-xmlfile==1.1.0
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.5
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
multidict==6.0.5
mysql-connector-python==9.0.0
numpy==1.26.4
openai==0.28.0
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
panda==0.3.1
pandas==2.2.2
parsel==1.9.1
pdfminer.six==20231228
pdfplumber==0.11.2
pillow==10.3.0
Protego==0.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
PyDispatcher==2.0.7
pymongo==4.7.2
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyMySQL==1.1.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
queuelib==1.7.0
requests==2.31.0
requests-file==2.1.0
Scrapy==2.11.2
selenium==4.19.0
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
tabulate==0.9.0
tldextract==5.1.2
tqdm==4.66.4
trio==0.25.0
trio-websocket==0.11.1
Twisted==24.3.0

typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
w3lib==2.1.2
webdriver-manager==4.0.1
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
zope.interface==6.4.post2
